+++ app/about_us/OurCrew.js
@@ -1,0 +1,11 @@
+const OurCrew = () => {
+  // TASK - React 1 week 1
+  // Create the "Our Crew section"
+  // Use the descriptions provided in /app/about_us/README.md
+  // Use the pictures from /public/crew
+  // Some inspiration ideas found in /data/inspiration_about_us
+  return (
+    <p> ADD OUR CREW HERE </p>
+  );
+}
+export default OurCrew;
