+++ app/about_us/OurValues.js
@@ -1,0 +1,10 @@
+const OurValues = () => {
+  // TASK - React 1 week 1
+  // Create the "Our Values" section
+  // Use the descriptions provided in /app/about_us/README.md
+  // Some inspiration ideas found in /data/inspiration_about_us
+  return (
+    <p> ADD OUR VALUES HERE </p>
+  );
+};
+export default OurValues;
