+++ app/about_us/page.js
@@ -1,41 +1,18 @@
 import styles from './page.module.css';
+import OurValues from './OurValues';
+import OurCrew from './OurCrew';
+import OurPartners from './OurPartners';
 
-// TASK - React 1 week 1
+
 // After you are finished with creating the page, move the OurValues, OurCrew, OurPartners components into their own files
 // OurValues.js, OurCrew.js, OurPartners.js should live in this folder
 // import and use the components from the newly created files
 
-const OurValues = () => {
-  // TASK - React 1 week 1
-  // Create the "Our Values" section
-  // Use the descriptions provided in /app/about_us/README.md
-  // Some inspiration ideas found in /data/inspiration_about_us
-  return (
-    <p> ADD OUR VALUES HERE </p>
-  );
-};
 
-const OurCrew = () => {
-  // TASK - React 1 week 1
-  // Create the "Our Crew section"
-  // Use the descriptions provided in /app/about_us/README.md
-  // Use the pictures from /public/crew
-  // Some inspiration ideas found in /data/inspiration_about_us
-  return (
-    <p> ADD OUR CREW HERE </p>
-  );
-}
 
-const OurPartners = () => {
-  // TASK - React 1 week 1
-  // Create the "Our Crew section"
-  // Use the descriptions provided in /app/about_us/README.md
-  // Use the pictures from /public/crew
-  // Some inspiration ideas found in /data/inspiration_about_us
-  return (
-    <p> ADD OUR Partners HERE </p>
-  );
-}
+
+
+
 
 
 export const Crew = () => {
@@ -51,9 +28,14 @@ export const Crew = () => {
           <h2>The crew</h2>
           <OurCrew/>
         </section>
+        <section className='card'>
+          <h2> Our Partners</h2>
+          <OurPartners/>
+        </section>
 
-         {/* TASK - React 1 week 1 */}
+         
          {/* Add in the "OurPartners" component here */}
+
       </main>
     </div>
   );
