+++ app/layout.js
@@ -17,8 +17,9 @@ export const RootLayout = ({ children }) => {
       <body className={inter.className}>
         <Navbar />
         {children}
-        {/* TASK - React 1 week 1 */}
+        
         {/* Import and use the Footer component here */}
+        <Footer />
         {/* Footer found in the ui/Footer.js folder */}
       </body>
     </html>
