+++ components/ui/Footer.js
@@ -1,15 +1,18 @@
-"use client"
+"use client";
 
-import { usePathname } from 'next/navigation';
-import styles from './Footer.module.css';
+import { usePathname } from "next/navigation";
+import styles from "./Footer.module.css";
 
 export const Footer = () => {
-  const path = usePathname().split('?')[0];
+  const path = usePathname().split("?")[0];
   return (
     <footer className={path !== "/" ? styles.footer : styles.hidden}>
       <div className={styles.footerDescription}>
         <h3>Galactica</h3>
-        <p>Explore the universe and beyond. Your journey to the stars starts here.</p>
+        <p>
+          Explore the universe and beyond. Your journey to the stars starts
+          here.
+        </p>
         <p>&copy; 2024 Galactica. All rights reserved.</p>
       </div>
       {/* TASK - React 1 week 2 */}
@@ -24,8 +27,6 @@ export const Footer = () => {
       </div> */}
       {/* Docs for the Link: https://nextjs.org/docs/pages/api-reference/components/link */}
 
-      {/* TASK - React 1 week 1 */}
-      {/* Add a new list item for LINKEDIN */}
       <div className={styles.footerLinks}>
         <h3>Follow us</h3>
         <ul className={styles.footerList}>
@@ -41,6 +42,9 @@ export const Footer = () => {
           <li>
             <a href="https://google.com">On the streets at night</a>
           </li>
+          <li>
+            <a href="https://linkedin.com"> linkedin</a>
+          </li>
           {/* TASK - React 1 week 2 */}
           {/* Create a <SocialMediaItem /> component and replace all of the list items! */}
           {/* it should accept the following props */}
@@ -50,4 +54,4 @@ export const Footer = () => {
       </div>
     </footer>
   );
-}
+};
